bunch of test stuff. ShittySituationsDog.h contains code for the newest patch to summon the interior props for the Higgins Helitours business and the Smoke On the Water Interiors.

+++ GTAVInfrastructure/World/Ambience/Weather.h
@@ -27,7 +27,7 @@ public:
 	const char* GetWeatherFromHash(unsigned long hsh);
 
 private:
-	std::map<unsigned long, const char*> WeatherTypes;
+	std::map<long long, const char*> WeatherTypes;
 	CWeather m_pWeather, *m_pComboWeather;
 	float m_fMixPercentage = 0.0f;
 };
