+++ LegitProject1/EndTypes/Marker.cpp
@@ -5,5 +5,5 @@
 
 void CMarker::Render(bool bob /*= false*/, bool faceCam /*= true*/, bool drawOnEnter /*= false*/)
 {
-	GRAPHICS::DRAW_MARKER(GetMarkerType(), Pos().GetX(), Pos().GetY(), Pos().GetZ(), Dir().GetX(), Dir().GetY(), Dir().GetZ(), Rot().GetX(), Rot().GetY(), Rot().GetZ(), Scale().GetX(), Scale().GetY(), Scale().GetZ(), Col().GetR(), Col().GetG(), Col().GetB(), Col().GetA(), bob, faceCam, 0, faceCam, GetTextureDict(), GetTextureName(), drawOnEnter);
+	GRAPHICS::DRAW_MARKER(GetMarkerType(), Pos().GetX(), Pos().GetY(), Pos().GetZ(), Dir().GetX(), Dir().GetY(), Dir().GetZ(), Rot().GetX(), Rot().GetY(), Rot().GetZ(), Scale().GetX(), Scale().GetY(), Scale().GetZ(), (int)Col().GetR(), (int)Col().GetG(), (int)Col().GetB(), (int)Col().GetA(), bob, faceCam, 0, faceCam, GetTextureDict(), GetTextureName(), drawOnEnter);
 }
